Implement stock filtering and add new sale registration interface

+++ src/main/java/org/mdv/dto/VentaDetalleRequest.java
@@ -7,6 +7,6 @@ public record VentaDetalleRequest(
         BigDecimal precioUnitario,
         BigDecimal totalLinea,
         String codProd,
-        Long ventaId
+        int ventaId
 ) {
 }
